application restructured
